requirements.txt
Flask==2.3.3
flask-cors==4.0.0
openai==0.28.1
numpy==1.24.3
tiktoken==0.5.1
PyPDF2==3.0.1
python-docx==0.8.11